# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      image_url:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        maxLength: 8
        minLength: 5
        type: string
      role_id:
        type: string
    required:
    - name
    - password
    type: object
...
  dto.EmptySuccessResponseWrapper:
    properties:
      message:
        example: success
        type: string
      status:
        example: true
        type: boolean
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
      message:
        example: failed request
        type: string
      status:
        example: false
        type: boolean
    type: object
...
  dto.PermissionResponse:
    properties:
      endpoint:
        type: string
      id:
        type: string
      method:
        type: string
      module:
        type: string
      name:
        type: string
    type: object
...
  dto.RoleResponse:
    properties:
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/dto.PermissionResponse'
        type: array
    type: object
...
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      image_url:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      role_id:
        type: string
    required:
    - name
    type: object
  dto.UpdateUserStatus:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/dto.RoleResponse'
      status:
        type: string
    type: object
  dto.UserResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/dto.UserResponse'
      message:
        example: success get user
        type: string
      status:
        example: true
        type: boolean
    type: object
  dto.UsersResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      message:
        example: success get users
        type: string
      meta: {}
      status:
        example: true
        type: boolean
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@simponi.app
    name: Simponi Support
  description: |-
    REST API for Simponi - multi-store inventory and omnichannel management system.
    This API supports store management, product management, external product integration, vendor management, and role-based access control (RBAC).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://simponi.app/terms/
  title: Simponi Backend API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: 'Get paginated users (Requires permission: GetUsers)'
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.UsersResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get list of users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: 'Create a new user (Requires permission: CreateUser)'
      parameters:
      - description: Create user request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/dto.UserResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - Insufficient permission
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict - User already exists
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new user
      tags:
      - Users
  /users/{user_id}:
    delete:
      consumes:
      - application/json
      description: 'Delete user by ID (Requires permission: DeleteUserByUserID)'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.EmptySuccessResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: 'Get user detail by user ID (Requires permission: GetUserByUserID)'
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.UserResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: 'Update user data by ID (Requires permission: UpdateUserByUserID)'
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - description: Update user request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.UserResponseWrapper'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
  /users/{user_id}/status:
    patch:
      consumes:
      - application/json
      description: 'Update user status (active/inactive) (Requires permission: UpdateUserStatusByUserID)'
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - description: Update status request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserStatus'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.UserResponseWrapper'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user status
      tags:
      - Users
  /users/profile:
    get:
      consumes:
      - application/json
      description: 'Get authenticated user profile (Requires permission: GetUserProfile)'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.UserResponseWrapper'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update current logged-in user profile
      parameters:
      - description: Update user profile request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Use JWT token with format: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...